+++ src/components/i18n/en.ts
@@ -1,0 +1,38 @@
+export const phrases = {
+  "webmapgl": {
+    "app": {
+      "menulabel": "Legend"
+    },
+    "basemap-switcher": {
+      "header": "Select a Basemap"
+    },
+    "basemaps": {
+      "title": "Change basemap"
+    },
+    "draw-toolbar": {
+      "cancel": "Cancel",
+      "confirm": "Continue",
+      "label": "Add a Feature"
+    },
+    "drawer-toggle": {
+      "label": "Toggle drawer"
+    },
+    "feature-add": {
+      "success": "Saved successfully.",
+      "failure": "An error occurred while saving.",
+      "label": "Add a Feature"
+    },
+    "field": {
+      "required": "%{field} is required."
+    },
+    "form-page": {
+      "back": "Back",
+      "submit": "Save",
+      "cancel": "Cancel"
+    },
+    "fullscreen": {
+      "enter": "Enter fullscreen mode",
+      "exit": "Exit fullscreen mode"
+    }
+  }
+};