+++ src/components/i18n/i18n.ts
@@ -1,17 +1,19 @@
-function getLang(defaultLang: string) {
-  if (!document) return defaultLang;
-  return document.querySelector('html').lang || defaultLang;
-}
+import Polyglot from 'node-polyglot';
 
-// TODO: Figure out how to load only the required translation.
-const lang = getLang('en');
-const translation = {
-  'es': {
-  },
-  'fr': {
-  }
-}[lang] || {};
+import { phrases as enPhrases } from './en';
+const phraseMap = {
+  en: enPhrases
+};
 
-export function _t(text: string) : string {
-  return translation[text] || text;
+(window as any).polyglot =
+  (window as any).polyglot || new Polyglot({locale: 'en'});
+const polyglot = (window as any).polyglot;
+const locale = polyglot.locale();
+
+polyglot.extend(enPhrases);
+if (locale !== 'en') polyglot.extend(phraseMap[locale] || {});
+
+export function _t(phrase: string,
+    interpolationOptions?: Polyglot.InterpolationOptions) : string {
+  return polyglot.t(phrase, interpolationOptions);
 }