+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added header facet widget.
 - Added facet class to form page.
 - Block click actions on disabled buttons.
+- Use polyglot for translation.
 
 # [0.3.1] - 2018-06-08
 - Prevent form submission if there are no fields in the page.